Nope, I could do it in Spiral due to its structural polymorphism, but doing it in F# is just too dangerous.
By dangerous I mean that responsibility should be separated evenly so one function is responsible for only one thing.

If I squished the passes together that principle would be violated. I very frequently change Spiral's code and having to modify the same thing in two places would be a natural way to introduce errors.

/u/combinatorylogic was right in that in ML it is a pain to do AST traversals. Spiral will be much more flexible in that regard.

So yay, I am done with with making improvements to the langauge.

It is time for play. I had enough of work.