I changed my mind.
What I wrote in the last post was right, but now that I am timing renaming more closely I see that it is not a 50% hotspot. It is a 90% one!

My conclussion is that it takes it 3.3s to rename maybe a total of 4k variables in the newest example. Out of that, 3.3s, 2.85 is how long it spends in the renamer function.

This is waaaaay over the line. Should I try ditching immutable maps?

Let me try running it through the debugger.

I want to understand the path it is taking better.