As soon as I stepped into the debugger I realized where I went wrong.
Oh no, oh no, oh no! I made a trully horrific error in the renamer. And this explains why the damn hash took so long! It explains why the hotspot just moved to a different place when I did this!

I understand it now. It cannot be anything other than the fact that it is renaming the same things constantly.