"This should do it.
I am going to have to do the same scheme for free variables.

I am going to use that elegant memoize function instead of this crap, but let me run it as a proof of concept.

Time for parse: 00:00:00.1110114
Time for prepass: 00:00:00.0279208
Time for peval was: 00:00:00.5792154
Time for optimization pass was: 00:00:00.0432932
Copied the code to: C:\Users\Marko\Documents\Visual Studio 2015\Projects\ConsoleApplication4\ConsoleApplication4\Program.fs
Time for codegen was: 00:00:00.0683528
Total renaming time = 00:00:00.0401806
See this. The renaming time was cut from 2.3s (I am using test34) to 0.04s! And I am betting the other hotspot is calculation of free variables now.

I did it! I diiiiiiiiiiiid it!!!

Let me do it elegantly now!

Ah, forget it, let me just do it normally. It will be faster anyway without unnecessary indirection.

7:10pm. Let me get rid of the optimization pass. It is such a bother.

Time for parse: 00:00:00.1109165
Time for prepass: 00:00:00.0283972
Time for peval was: 00:00:00.4345861
Copied the code to: C:\Users\Marko\Documents\Visual Studio 2015\Projects\ConsoleApplication4\ConsoleApplication4\Program.fs
Time for codegen was: 00:00:00.0662714
Total renaming time = 00:00:00.0329384
This is how it looks like now.

Damn it, I am such a fool. Trully such a fool.
Let me try scaling the n again.

7:15pm. Yeah, now I can do 160 in 2.6s.

Let me see what code gets printed and I will try profiling it again. I will take this shitty lesson to heart. I think that given how long it takes it to loop 50k times, I think that 10k lines of code generated in 1-2s would be an achievable goal for the pevaller.

Ok, with n = 160 it compiles to 496 methods and 11.2k lines of code. For the amount of time that it took, I feel this is a much more realistic number.

With this single change, I managed to improve the timing by orders of magnitude.

Let me compare it with the new libary to see if there are any differences now.

Also at 11.2k LOCs it does seem like VS is starting to struggle to do typechecking.

7:20pm. Sadly, it terms of timing the two libraries are neck to teck. There is no difference in performance. It just comes down to which style do I like more.

What I want to do is profile more. There is still some superlinear scaling as I vary the n."

Though at this point it might be simply because the number of variables is getting larger.

Let me try it again. This is so wonderful. I might be able to draw out the full power of this style after all.