No, I really can't tell at this point, though as a baseline, the fact that pevalling is within factor of 10x in speed compared to codegen is telling that the pevaller is not that slow at this point. And then the fact that codegen is scaling non-linearly is telling as well, that it might be code size at this point.
It might be worth trying to convert to CPS as an experiment, but the profiling story is not consistent with the GC being the bottleneck.

I should try it...there isn't really much else I can do at this point. I'll sleep on it.

I think 9pm is a decent time to stop programming for the day. I really had a lot difficulty starting, but I got damn into it today. I had not expected that I would resolve this big issue that has been plaguing me.

I'll try moving the pevaller to CPS tomorrow.