I managed to push the time down to 6.6 by doing another little optimization and switching to the new library.
This is quite good. I was wrong that I could optimize the env function further, but really at this point I think I have the whole compilation speed thing under wraps.

I am satisfied.

That having said, it is not compiling the hacker_rank_2 problem anymore so I will save here and backtrack in order to make sure that the changes I've introduced really were breaking.

The program typechecks just fine, but for some reason a piece of code is missing. How strange.