I dissed the `let inline` declaration in that issue I opened, but now I see that it is fantastic. It is especially effective on functions taking other higher order functions as argument.
Honestly, when I wrote my last commit I had succeeded in bringing down the time from 10s, which was already a vast improvement on what I had even yesterday before my epiphany and I thought I had the speed under wraps.

However, when I tried adding inline on function taking higher order arguments, the speed just exploded so I kept doing and doing it, purgning out the inefficiencies.

As a result, the test34 now finishes in 4.05s.

This is really, really amazing. It is downright unbelievable.

The only other source of inefficiency remaining in the pevaller would be the active patterns in the apply function. I already redesigned it so it does not take a higher order function as argument, but I suppose there is some room left to go.

Let me save here and I will do one last push.

This is just too good. It is too good to give up.