No, it is not working. It is just too tedious without active patterns. I need Spiral to properly optimize this. I NEED the active patterns as much as I need higher order functions for the sake of keeping order in my codebase.
Performance be damned here. Let me take a break.

I certainly deserve it after the heroic efforts of the past few days. There is no reason in the slightest to get anxious.

With this I absoultely have enough.

Tests 34 and 36 are the absolute worst case stress tests for the style I am trying to draw out. To generate 22.4k lines of code in 4s, yeah I can live with this.

With this I can do it, I can finish the library.

From here on out, I think I will give these dumb tests a rest and start doing more useful things.

The parser library needs more functions and one thing in particular as an offshot of that which I need is the print function. I also need proper error handling in my DSLs.

It is amazing how much plumbing even the simple problems will need.

...Yes, no matter how I look at it, this should be enough as the basis. Everything I said before I figured out the memoization trick and applies 100 times more now.

One thing I also need to do before call it a day is open an issue on the F# repo to tell them that a mutually recursive block of 22.4k lines crushes the IDE.

As the last thing, let me profile the program I want to see if the memory consumtion went down.