"5:10pm. Done with the bath. Let me adjust test36 so it sums the integers at the end and then I will show to the F# guys.
...Yeah, I probably should split the env into individual arguments as an experiment, but forget that for now.

inl ret =
    inl on_succ x = Tuple.foldl (+) 0 x
    inl on_fail x = 0
    inl on_fatal_fail x = 0
    inl on_type = int64
    module (on_succ,on_fail,on_fatal_fail,on_type)

Now I wonder what I will get.

5:20pm. I think I will just post the 13.4k lines one on the F# sub.

5:25pm. It grinds the Preview tab to a halt too. I am going to modify the output so I get the thing in the same folder.

5:40pm. Holy crap the thing is really 10MB? This is the 26.8k lines one with n = 320.

Actually I never realized how huge it is until I opened it in the browser and had all the arguments word wrapped.

Let me try pasting it in Writer so I can get a sense of how many pages it is.

2383 pages. Yeah, forget about it. I'll get laughed out of the repo. Now that I am writting this out, I am realizing how absurd asking whether the IDE should slow is.

For HM inference, type annotations do not matter and given that, it has the responsibility of having to keep track of...upwards of 50k variables. And probably a few multiples of that.
I should probably start studying the .NET IL. I already read that book so I have some idea how it works.

As an aside, it seems the spellchecker for Writer is broken again. I really need a better editor."