While I did a good job on improving performance yesterday, it occured to me during the night that I also made a great big mess out of the codebase. What I am going to do now is clean the thing up a bit.
What I am refering to is specifically the way I've nodified that Dictionary and the HashSet. Nodify really is not meant to be used like that and I was uncomfortable the whole way as I was doing it.

But even before that, putting those used variables and refs directly in the TyJoinPoint is a mistake. I will make a dictionary specifically for join points and move the relevant stuff that.

Putting everything in a single function was a good idea in the end as it allow me to be flexible and take advantage of lexical scope without having do pass around global records.

Now let me go a bit further with this.