Given that it compiles test36 without type error, I have no problem believing that my refactor is correct.
This redesign quite nicely simplifies the whole thing.

I also realized that eval_renaming used a higher order function and inlining it seems to have produced a 5% speed up. Nice.

I guess I'll have to deal with arrays, but forget them for now.

Unlike the extreme effort of last month that it took me to finally get pattern matching to work and then do the .NET integration and debug the whole thing enough to have it compile code, and then the extreme effort of the past 2.5 weeks where I speed up the compiler thousands of times, doing arrays properly won't require any conceptual breakthroughs nor much effort.

In the parser library, I need to add proper error message handling and besides that, I really need to implement sprintf. I should also implement its opposite - the sreadf for those shitty HackerRank problems.

Before I take off for a bit, let me just say a few things that are left.

Looking at the present, I think that at this point, most of the major challenges for making the first version of the Spiral language have been surmounted.

I started by making a very basic macro language for writing Cuda code and evolved it into the beast of partial evaluation that I have here. Right now, by the virtue of its design, Spiral is the fastest, most expressive functional language in the world.

The rest that remains is to write the libraries for it that will actually make it useful for real world things, but that is just a matter of putting in the time.

Until I get the agents to give me a hand, I do not think I will have the same excitement of working on the language that I have had for the last six months.

The language has evolved remarkably. The things that it would need in V2 to make it palatable to non-humans like being able to fuse the evaluator into the libraries is something that I won't even attempt on my own until I attain a Channel. There would be great value in V2 of the language, just not to me while I am on my own.

To prove that the basic connection to reality that the language gives me is sound, I will resume the old quest of trying to make some cold, hard cash.

I'll do it at a leisurely pace as I always have. It will take a few months. Hopefully by the time the new year rolls around and it is time for another update, I will have something worth presenting.

For today, how about I make it my goal to do another HackerRank problem and then I will move from there?