I think that with this I have a decent way of handling modules in a top down fashion.
I am done for the day, I had enough.

Today, I wanted to start on doing sprintf, but as can be seen, building up infrastructure to that took me a while.

Writing the parse_modules function in CPS was quite an interesting challenge. It definitely took me a while. Getting better at CPS is merely a matter of practice.

The strange type errors when I program in that style in F# are definitely distracting. For an hour today, when I tried to get the grip on the function most of what was in my head was static and it was like I completely forgot how to do CPS, but at some point I realized how straightforward it was and I completed it without a hitch.

I think overall that despite some slow parts, I was in the zone plenty today and got some real work done. Each and every day is another step closer to Transcedence.