"8:25am. I had a great idea last night. I figured out how to do heap records. Also I should add something like a CSE switch to destructure via an argument instead of a global flag. That will ensure what while the string is in scope, it will get rewritten properly.
It won't be at all difficult to do and the best part is that allocating variables on the heap would relieve the renaming pressure. I wouldn't have to track the variables at all. I should take advantage of that to put those pure modules inside records so the renamer does not have to iterate over them.

8:45am. I need to do some more thinking before I commit. Like how would I do shared variables then. If I had a record of heap tuples with shared variables, how would I ensure they got shared?

Well in nested heaps there is not much that I can do, but maybe during construction if a heap is inside a heap, I can just get rid of inner ones and treat them as stack allocated?

Indeed such a thing would work, but I would rather prefer to be explicit in this case.

I think a basic way of doing this that composes with tuples and modules would be the best way to do this.

I guess I was wrong about Spiral not needing any more features.

But I do not need this stuff right now. I am happy enough to leave it for later."

Right now I am just getting distracted so I guess I will start. Let's see if I can put a dent into the sprintf function.