The code works, but there is too much of it in the new library. This is quite problematic as the code size not only went up by 50%, so did the compile times.
No, I said that there would not be any great challenges in the future, but I see now that was just wistful thinking. And it seems I am not done with performance debugging either. I am going to have to look for ways of squeezing the code size down.

I am going to have to compare this library for the previous one so let me use this commit as a save point.

I will finish sprintf and then figure out what to do with the code size after that.