I think I will call it a day here. I honestly can't go on anymore. I will look into...ah, I just realized something.
Instead of storing the unique key in all of the Nodes, I can in fact just store the hash only instead and that would still be a remarkable improvement. That way I would not have to hold dictionaries filled with garbage. During renaming, there is a possibility of using reference equality anyway. As a matter of fact, merging equal nodes under the same banner is something I could do there.

But in the vast majority of cases they will either not be equal or have the same reference anyway.
Yeah, this is the correct solution, without a doubt.

I was super wrong yesterday. The heat of improvement is still there. There is so much basic stuff to be done.

I am going to try redesigning the pevaller tomorrow.

I will keep Expr Nodes as they are.

Apart from that, I need to figure out why passing the pos in the state caused the code size to baloon.

That will have to wait for tomorrow.

What I did today was trully complete the sprintf function. Adding things like newlines and tabs will not be a problem.