I managed to root out the code blowup issue and now it takes two methods to compile the pint64 and just one to compile the spaces and this scales linerarly as n is varied in test36. In fact the resulting code is even more concise than previously, but the compile times got longer by 10% unfortunately.
That increase can be explained by having extra function applications in a hot path so I won't worry about this. I did some redesigning so that is supposed to be expected.

I'll admit, I am surprised that the first thing I tried was enough to solve this issue.

Now what will I do for the rest of the day?