After thinking about it for a while it occurred to me that the pevaller is not invariant to argument positioning so I made test40 to confirm it.
It turns out that was the case. How strange.

This is a bug in the semantics of the pevaller and I will have to spend some time fixing it as it is causing a lot of code blowup. But after I am done passing state should be equivalent to what I had before.