The last two hours since I decided to take a break were very productive in terms of brainstorming.
I figured out how to extensible pattern matching which would allow me do things like matching on lists and arrays in the same vein F# does now, except I will be able to build them directly in the language.

Also, as I was thinking about arrays it occurred to me that thanks to the continuing improvements in the partial evaluator, I should just get rid of all the complicated machinery and build them in as a library as well.

What I am doing now where I am directly scanning for sizes does not make the most sense.

It is not that it never occurred to me before, it is just while I was designing how do arrays Spiral had nowhere as developed capabilities as it did now so I ended up mentally frozen on the old designs.

For the rest of the day, I think I will lighten up the arrays and finally call it a day.