Now that I had some time to deliberate, I know what my plan will be from here.
First off – errors. It seems simple at first glance, but is actually quite difficult. I have a nice idea on how to do it, but will really have to reshuffle the parser to get it to work. And in the end, I think that it is too early.

Right now the errors are ugly, but they are also convenient for myself while I am still debugging the language. They are not fit for users, but they are fit for me as a developer. Once it is time to do the introductions, I will make this last final change and hide the underlying machinery.

So I will leave error messages for last.

Then here is what I will do next. I will go forward with my plans to the extend the pattern matcher. I will rebind the `^` pattern to `!` and the use `@` for partial active patterns and `#` for extended active patterns.

Also I will make parse_n_ints consider the ints it is getting and switch to the more efficient closure using method for n > 5.

Apart from that I will extend the standard parse_ints so it takes a loader as an optional argument which like for Inline methods in Fparsec will allow me to tailor it to use arrays or ResizeArrays as I see fit.

I think that once I have all of the above, loading data for those HackerRank challenges won't be a problem and I will be able to play for a while. Besides that I will have to make the standard library functions for Array and List.

Ok, first off let me start by dealing with the parse_n_ints...

Also, I am going to have to hack it a little so it does not do what I mentioned except during testing.