I did a whole bunch of stuff since the last commit, including extending the pattern matcher with partial and extension patterns.
I changed the precedence for the cons and the or patterns and I did the parse_n_ints. I also made the number parsers a good bit safer.

There is a decent bit of genius in the way I deal with pattern matching in the language now, especially in the pattern compiler. For a partially evaluated language like Spiral, optimizing pattern matchers is like the best case scenario for it.

It also well suited for parsers when compiled to a platform with GC for the access to closures.

I think with this that I can seriously start work on an array library. After I do that and touch up parsers a bit more, I will be able to do any kind of HackerRank problem without being forced to put up any more infrastructure.

As usual, today I started off lethargic and then did a good bit of good work. I'll take off here.

It won't be long now. I kept extending and extending my work on the language and slowly, but surely, it is starting to approach completion.

At this rate, it is not unthinkable that I might return into doing ML by November.