Let me just say that I am glad I did not try to tackle making PatType yesterday because I am still perplexed as to how to do it even now after working on it for over an hour.
Now that I am done with breakfast and ready to resume, I think I will dedicate some time to thinking this through.

It comes down to the fact that I am essentially using the pattern compiler in an unhigienic manner and have no way to rebind the variable for the type check.

Agh, I do not feel like explaining it here any longer. Let me just focus on it for a while.