```                 let rec f pat' on_fail = function                     | PatAnd _ as pat -> op(ErrorType,[lit_string "And patterns are not allowed in extension patterns."]) |> pat_part_active a (pat' pat) on_fail                     | PatOr l -> List.foldBack (fun pat on_fail -> lazy f pat' on_fail pat) l on_fail |> force                     | PatCons l as pat -> vv [type_lit_create (LitString "cons"); l.Length |> int64 |> LitInt64 |> lit; arg] |> pat_part_active a (pat' pat) on_fail                     | PatTuple l as pat -> vv [type_lit_create (LitString "tup"); l.Length |> int64 |> LitInt64 |> lit; arg] |> pat_part_active a (pat' pat) on_fail                     | PatType (a,typ) -> f (fun a -> PatType(a,typ) |> pat') on_fail a                     | PatWhen (pat, e) -> f (fun pat -> PatWhen(pat,e) |> pat') on_fail pat                     | PatClauses _ -> failwith "Clauses should not appear inside other clauses."                     | pat -> vv [type_lit_create (LitString "var"); arg] |> pat_part_active a (pat' pat) on_fail                 f id on_fail pat ```
It took me quite a bit to figure out how to make the above work. Around 4.5 hours or so.

Phew, I am finally done with extension patterns. I had not expected that figuring out how to design the thing so the `type` pattern is supported would take me so damn long.

I almost don't feel like doing anything for the rest of the day, but I'll try to do a bit more.

Before this issue occupied me so much, what was the plan that I had for today?