I am going to remove the old parsing module from the library and get rid of the tests as well. Right now they are like an achnor that is holding me back from making necessary changes to the library.
When I am done I will redesign the parser so I can also include the userstate.

Also I can't fucking go through a hunded lines manually and change every little thing every time I want to extend the parser. I need to put in some thought on how to make it extensible from the start.

The answer for this are modules.

I need to add some better syntax for it.

None of this Tuple.upon' nonsense. I think I should reuse F#'s record syntax for this sort of task.

As a language designer, I always need to keep in mind that I need to operate in terms of what should be, not what I have.

It is time for another extension I think. Having extensible modules would be a great benefit to the language.