I've finished making the third version of the parsing library that uses modules. It took me a while that is for sure, but now that it is in this form not only is it more readable, but it also perfectly extensible as well.
The new version of the parser simply looks wonderful. This is how it was meant to be written. I absolutely must have this feature.

Right now the example still cannot be run and so it is time for the final stretch. I need to implement the module pattern and the module parsers. It might be wise just like for the patterns for the parser to return the module as an AST instead of trying to compile it directly.

That would certainly be for the best.

Before I start, let me check out how other languages handle modules. I think it is about time I do some research in what all the fuss is about the ML modules.

Well, I'll do this step by step. I certainly won't be able to finish all of this today, but maybe three or so should be enough if I had to guess.

Either way it will be worth it.