I want to continue, but there is no realistic way I can go on past this point. I implemented the first part of the new module features, namely the pattern matching aspects of them and tomorrow I am going to add the special syntax to construct them using the {} notation.
Besides that, I am also going to support the lens update notation as can be seen in the parer example.

So I am about a third done with this. If I am lucky I'll be able to finish the rest tomorrow and move on to testing the day after that, on Wednesday.

To be honest, today it does not feel at all like accomplished much. I do not understand how it happened, but I only really started working on the parsers proper past 2pm.

Rewriting the parsing library is the culprit. It seems easy, but writting in a language that does not exist yet is quite mentally exhausting. It took me quite a bit of time finish the leftover from yesterday.

I first noticed this when I wrote the Cuda kernels months ago in imaginary Spiral.

It is not so imaginary now. A bit by bit I am bridging it closer to reality.