Jesus, I am done with adding the lens patterns to both the pattern matcher and the core language. I haven't testerday it yet and will leave it for tomorrow as I am really tired at this point and have no stomach for a potentially tought debugging session.
It is time for me to rest.

For two days work, I definitely did quite a bit. No mistake, these new syntax extensions are a powerful added to the language's usability.

Just consider, everything I have here apart from (boxed) recursive types will also be usable on the GPU. This is something beyond amazing for a language with Spiral's feature set.

Looking deeper into the future, a few years hence, when the langauge is not usable just for myself, but for other people as well, I will have advanced the art of GPU programming by decades singlehandedly. Not just that, but a demonstration of an ultra-efficient functional language will also be a gamechanger for high perfomance oriented users.

What I did today was just another part of what will make Spiral worth using.