The end result of the performance testing is the the new parser is 20% slower to compile. I'd considered testing that 26k LOC program, but I am not sure how to compile it without crashing the IDE, so I'll give that a pass.
I also do not care about refactoring the parser so the return arguments are uncurried.

The reason why the new parser is slower is most likely because of all the extra checks in the module bindings. The previous version was undoubtedly more direct.

At any rate, 20% is a good tradeoff for this vastly increased flexibility. Last night it occured to me that as it is now, Spiral has become good at OO. Which is a pretty good thing.

The OO approach of Spiral is pretty nice. The way C# and Java do it has made me think that OO is worthless garbage, but what I have now is quite enough to make change my mind. The way Spiral does it with duck typing and composition is how it should be. It is mixes very well with the general functional approach and pattern matching, which is why I like it so much.

It is definitely a natural part of the language.

I think I deserve a break after the back breaking work of the last few days, but I feel like going on.

There is a thing I wanted to check out ever since I fixed the method call semantics - whether making the state dynamic reduces the code size. Last time I had the bug which lead to strange behavior, but now I think it should do what I want.

Let me check it out.