inl tuple l {d with state {ret with on_succ on_fail}} =     inl rec loop state l ret =         match l with         | x :: xs ->             x {d with
See how the loop fetches the d from the top level and that d still has the old pos. It should not have taken me this long to realize this, I think I was on the verge of it once or twice, but could not make the breakthrough.

Let me save here and I will take a look at the old branch.