I am done for the day. I fixed a few bugs, improved the compile times* and did what I set out to do today. Good work, me.
Turning tuple into tuple_template did make the times 5% worse again, but I am not going to loose sleep over that crappy test. It is meant as a rough guide anyway.

I should partially evaluate the whole parsing library by fusing it with the evaluator if I want to get rid of the fat anyway.

Damn I am tired. I think starting tomorrow, I will be able to move to doing HackerRank exercises and so probably the first thing I will do is build up the array functions.

The last week or so was really backbreaking work. Today was not easy either.

Debugging is swimming in a sea of perplexity. Debugging is what happens when you give it your best and fail. It is never fun for me, but at the end of it, you always learn something new.