Ok, I can confirm that the changes I did to sprintf yesterday were correct. I guess the next thing is to put printf back in again.
The thing next on my agenda would be to start work on the array sum problem. That means moving into making functions on arrays and lists.

After I have those two and after I've finished playing with the language for a bit, I can start work on putting in the Cuda backend and then start making stuff for the ML library.

Even though it is 2pm now, I haven't done a thing today due to bad weather. The storm literally did not stop until 1pm and I was just lounging in bed till then.

I won't bother trying to make up for lost time, instead I fully intend to take it easy today. I'll put in a few hours and call it a day.