I hear some rumbling in the distance, so the weather might be bad today. We'll see how it goes.
I've solved the Diagonal Difference problem, but unfortunately I do not have the assert implemented.

I'll put it into the pevaller right away. Actually, assert will be made in the host language, I'll just put in the more primitive failwith instead.

This does demonstrate one thing to me though, Spiral as it is now is not complete after all.

I really need monads or some kind of effect handling system. CPS is more like a functional anti-pattern as good as it is. It is not really fit for humans.

Honestly though, there is no way I am doing that one. Effect systems are experimental right now, and monads would need an advanced type system with higher kinded types. No way I am implementing that stuff.

If Spiral ever gets popular and the users want more advanced features, they can pay me to do it, otherwise I will just use what I have here and not fuss about it too much.

This is not like pattern matching which I knew I would be able to implement with some effort.

Implementing an effect system would be worth a PhD. If I had to do it...it would be when I've made at least a cool million and not have to worry about ending up on the street if things do not pan out.

...The weather is turning bad again. I think I will put of programmign until the storm passes.

Putting in a simple exception won't be too difficult at any rate..