On 3 out of 10 tests on the HackerRank site, I get segfaults.
I can't be bothered to actually buy the test case, but there is one thing I need to check. That being that that recursive loops get optimized properly.

It might be the case that with the way I am writing out by first assigning them to statements inhibits tail recursion optimization causing the loops to run out of stack space.

Let me make a little test for this.