It seems my intuition was right on the money this time. Now that I optimized away those unecessary statements and the dead weight unit parenths, the generated code looks a decent bit cleaner as well.
Great, great.

In the discussion page seeing just how much clearer the C++ solution is, is really warming me up towards doing effect handlers.

Right now, while highly expressive when written in CPS, the language is quite unwieldy to use because all the parameters that need to be passed around manually. This also leads to errors and long refactoring times.

With staged effect handlers, Spiral would really exceed the MLs in expressiveness. Just consider it, having exceptions on the GPU would absolutely break people's minds!

And these parser combinators would be absolutely trivial to make and extend in language with such a feature.

With a single move, I could push the language beyond the state of the art!

And the unidirectional type propagation that I am using is in fact possibly the only scheme that would have a chance of fully typing the handlers.

The time is not quite right that I jump into this feature, but it definitely is the point where I should seriously consider studying it.

With that, let me do the `https://www.hackerrank.com/challenges/birthday-cake-candles` problem. Doing just the first one took me forever, but my mind is wandering due to getting excited about the possibilities of effect handlers.