Done with the Birthday Cake problem. It was a simple fold, but I had to implement the elif statement for it which was easy. I also had to add the getting of fields from .NET types so I can get the Int64.MinValue.
Tomorrow I will take a break from this to do some imagination programming again. I am going to try and imagine and rewrite the parser as if I had effect handlers.

I am not sure whether or not they are worth it, so I need to program with them for a bit in order to get a feel for them.

This sort of programming with features that I do not have yet seems to be what I am doing more and more as I grow as a programmer.

I also want to do some research on Eff and read that paper on their implementation in C properly.

Right now I am too tired to continue, so this will be my last commit for the day. I did two problems today in the end.