I can confirm that the inm statement works as intended. The inmp function turned out to be a simple one liner and worked the first time. Just a while ago, I saw a scheme tutorial where it looked quite a bit harder with the macros and all.
Oh well, this is good all in all.

Now...I was really into it when I did the rewrite, but now I do not feel like it. Nevertheless, let me put in a bit of effort and hopefully as usual I will get into it. That is the case usually.

I'll give it a shot.