I spent the last two hours trying to fix a segmentation fault (stack overflow) only to conclude that the fault lies on the HackerRank side.
It is true that I had to modify the compiler so that no useless intermediaries are created in if statements. This was enough to fix the overflow on my side, but testing it on Ideone it seems to pass. On `Repl.It` it seems to run out of memory at roughly the same time it does with HackerRank, at about 15-17 iterations.

I made a feature request for tail call optimization, but I doubt it is going to get those guys to move.

The most confusing thing is that I did manage to beat it with the previous iteration of the library.

How strange. Let me try the array sum problem again and I will call it a day.

Actually, I should also test compilation speeds on the test I've been using. I'll do that promptly.