Wow, I am tired. I started a bit latter than usual today, but I really kept at it.
With this I am done with the parser debugging. Apart from sprintf, which I'll have to make a note to test tomorrow, all the regular tests pass.

Most of my time today and yesterday was taken by studying the effect on various monadic patterns on the generated code.

The runtime performance will be identical to the applicative style since everything gets nicely inlined, but the compile time performance takes a hit.

I estimate that the overhead of using monadic syntax is at ~2x for compilation time. I think this is quite a decent tradeoff.

Unlike in Haskell where there is a significant runtime penatly, monads can be used in Spiral without real difficulty.

The primary problem is when they cause the pevaller to start diverging as it become more difficult to understand to know what goes where, but I am guessing that getting better at this will come with experience.

I speculate that these compilation divergence will act as a filter for keeping out bad code.

Sure it is bad that the thing won't compile, but this would be far preferable that in Haskell where you would get memory leaks and runtime hits.

Now...

HackerRank is annoying me. I do not trust it anymore. I am going to have to find an alternative site to practice on, preferably something that allow me to use functional code with tail call optimization.

It is kind of ironic that after modifying the codegen twice so that this does not happen, the submission that passed previously stopped working. This pisses me off.

Also I can't figure out why the fold function gets duplicated in test55.

Along with testing sprintf, I'll save that for tomorrow.

It won't be long now. A few more week of this, and I will fully grip the reigns on the new language.