I figured out what the type error is. Passing d from the previous step was a pretty bad idea. I removed the state_d function having learned my lesson.
Now it is all set. I'll remove the old parsing library and then finally move on to doing programming challenges for a while. Maybe a week or two. This is just so I can get a feel for the language. After that I'll bring in the Cuda backend and deal with the ML library.

Let me deal with phasing out parsing3 to start things off.