Wow, I have been stuck on this for a while. I did some good thinking thanks to it.
The end conclussion of all my thinking is that while I could use a tuple of queues representation, I have no way of representing the damn parent pointer.

I have no way of passing innate Spiral discriminated union types to the .NET side. Without that I won't be able to get the trace. This has made me realize that good integration will be harder than it seems.

There is a schism between what the codegen is capable of using built in types and what the pevaller is capable.

...This is so difficult. Good integration is difficult. It is hard enough to have the language be consistent internally without the platform it is on making a mess of things.

This also means that if I can't make full use of the .NET ecosystem, the benefits of being in it drops down in value accordingly.

I'd definitely get more speed if I compiled to LLVM for example.

But forget this for now. So I can't make full use of generic types then.

What I did figure out in the past 3 hours though is how to do breadth first search in a functional manner. Last year while I was studying Haskell for the sake of contributing to Futhark I could not quite internalize it.

But now I understand in 100%, just like Y Combinators.

It won't be hard to do at all thanks to arrays. The only thing I need to do is just expand and then flatten them.

Let me do so now.

I hoped I would do more than one problem per day, but it seems the circumstances are not on my side. Making a mainstream language is a huge effort.

By the time I am done, I am probably going to build Dictionaries and HashSets into the language or implement them directly.

The only thing on my side is that I do not need to do all this stuff.

I won't hang around doing these things too long. Let's say that 10 days is the limit.

After that I'll get into the action properly.